[issue449] basic windows support

+++ src/search/landmarks/exploration.cc
@@ -166,10 +166,10 @@ void Exploration::build_unary_operators(const Operator &op) {
 class hash_unary_operator {
 public:
     size_t operator()(const pair<vector<ExProposition *>, ExProposition *> &key) const {
-        unsigned long hash_value = reinterpret_cast<unsigned long>(key.second);
+        uintptr_t hash_value = reinterpret_cast<uintptr_t>(key.second);
         const vector<ExProposition *> &vec = key.first;
         for (int i = 0; i < vec.size(); i++)
-            hash_value = 17 * hash_value + reinterpret_cast<unsigned long>(vec[i]);
+            hash_value = 17 * hash_value + reinterpret_cast<uintptr_t>(vec[i]);
         return size_t(hash_value);
     }
 };
