+++ src/search/landmarks/exploration.h
@@ -10,6 +10,7 @@
 #include <ext/hash_set>
 #include <ext/hash_map>
 #include <cassert>
+#include <stdint.h>
 
 class Operator;
 class State;
@@ -81,7 +82,7 @@ struct ExUnaryOperator {
 
 struct ex_hash_operator_ptr {
     size_t operator()(const Operator *key) const {
-        return reinterpret_cast<unsigned long>(key);
+        return reinterpret_cast<uintptr_t>(key);
     }
 };
